move client/unversioned/fake to client/restclient/fake
