+++ pkg/kubectl/BUILD
@@ -157,8 +157,8 @@ go_test(
         "//pkg/client/clientset_generated/internalclientset/typed/core/unversioned:go_default_library",
         "//pkg/client/clientset_generated/internalclientset/typed/extensions/unversioned:go_default_library",
         "//pkg/client/restclient:go_default_library",
+        "//pkg/client/restclient/fake:go_default_library",
         "//pkg/client/testing/core:go_default_library",
-        "//pkg/client/unversioned/fake:go_default_library",
         "//pkg/controller/deployment/util:go_default_library",
         "//pkg/kubectl/testing:go_default_library",
         "//pkg/runtime:go_default_library",
