+++ pkg/kubectl/cmd/testing/BUILD
@@ -23,8 +23,8 @@ go_library(
         "//pkg/apimachinery/registered:go_default_library",
         "//pkg/client/clientset_generated/internalclientset:go_default_library",
         "//pkg/client/restclient:go_default_library",
+        "//pkg/client/restclient/fake:go_default_library",
         "//pkg/client/typed/discovery:go_default_library",
-        "//pkg/client/unversioned/fake:go_default_library",
         "//pkg/kubectl:go_default_library",
         "//pkg/kubectl/cmd/util:go_default_library",
         "//pkg/kubectl/resource:go_default_library",
